# packages/protocol/openapi.yaml
openapi: 3.1.0
info:
  title: ClickClack API
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
paths:
...
  /api/auth/magic/request:
    post:
      operationId: requestMagicLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestMagicLinkRequest"
      responses:
        "201":
          description: Created local magic-link token in dev-bootstrap mode
        "403":
          description: Dev-bootstrap token minting requires a loopback client and local request host
        "501":
          description: Magic-link delivery is not configured
  /api/auth/magic/consume:
    post:
      operationId: consumeMagicLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumeMagicLinkRequest"
      responses:
        "200":
          description: Created session
  /api/auth/password/login:
    post:
      operationId: passwordLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordLoginRequest"
      responses:
        "200":
          description: Created session
        "400":
          description: Missing identifier or password
        "401":
          description: Unknown identifier, wrong password, or an account without a password set
        "403":
          description: Cross-site login requests are not allowed
        "415":
          description: Content-Type must be application/json
        "429":
          description: Too many attempts for this client address or identifier
        "501":
          description: Password login is not configured
...
components:
...
  schemas:
...
    RequestMagicLinkRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        display_name:
          type: string
    ConsumeMagicLinkRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
    PasswordLoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Account email address or handle
        password:
          type: string
...